Support for building every commit
This basically works by holding back commits until GoCD can
schedule them. This assumes that GoCD starts scheduling the given
commits right after the plugin gives them to the server. This
will prevent the case where GoCD is building one commit whichs
takes long time to build. Meanwhile other commits would be given
to GoCD and it would merge these commits. That would result some
of the commits to be skipped.

This is useful when using tools like Gerrit which should make
GoCD to build every change set and report the result back to Gerrit.
Without this feature GoCD will skip some changeset if the changeset
are related to each other.

This feature requires access to Go's pipeline API so that it can
check if the pipeline can be scheduled.

This will also require the pipeline to have only one stage. Otherwise
GoCD (tested on versions 16.2, 16.6) may require manual triggering of
the second stage even when pipeline isn't configured to behave like
this.