Fallback to detecting latest revision if revisions-since call fails
This usually happens when an attempt is made to force-push, and in the
meanwhile, the flyweight folder is deleted. The plugin tends to perform
a `git log last-commit..HEAD` and the `last-commit` is no longer in the
git object database, thereby rendering the plugin useless.

A workaround is to ask that the old commit be force-pushed (hopefully
whoever pushed the old commit still has it in their reflog)